+++ src/facet/data/_sample.py
@@ -6,6 +6,7 @@ import logging
 from copy import copy
 from typing import Any, Collection, Iterable, List, Optional, Sequence, Set, Union
 
+import numpy as np
 import pandas as pd
 
 from pytools.api import AllTracker, to_list, to_set
@@ -71,7 +72,8 @@ class Sample:
     ) -> None:
         """
         :param observations: a table of observational data;
-            each row represents one observation
+            each row represents one observation, names of all used columns must be
+            strings
         :param target_name: the name of the column representing the target
             variable; or an iterable of names representing multiple targets
         :param feature_names: optional iterable of strings naming the columns that
@@ -81,28 +83,13 @@ class Sample:
             observation
         """
 
-        def _ensure_columns_exist(column_type: str, columns: List[str]):
-            # check if all provided feature names actually exist in the observations df
-            available_columns: pd.Index = observations.columns
-            missing_columns = {
-                name for name in columns if name not in available_columns
-            }
-            if missing_columns:
-                raise KeyError(
-                    f"observations table is missing {column_type} columns "
-                    f"{missing_columns}"
-                )
-
         # check that the observations are valid
-
         if observations is None or not isinstance(observations, pd.DataFrame):
             raise ValueError("arg observations is not a DataFrame")
 
-        observations_index = observations.index
-
-        if observations_index.nlevels != 1:
+        if observations.index.nlevels != 1:
             raise ValueError(
-                f"index of arg observations has {observations_index.nlevels} levels, "
+                f"index of arg observations has {observations.index.nlevels} levels, "
                 "but is required to have 1 level"
             )
 
@@ -111,7 +98,7 @@ class Sample:
         targets_list: List[str] = to_list(
             target_name, element_type=str, arg_name="target_name"
         )
-        _ensure_columns_exist(column_type="target", columns=targets_list)
+        _ensure_columns_exist(observations, column_type="target", columns=targets_list)
 
         self._target_names = targets_list
 
@@ -142,7 +129,9 @@ class Sample:
             features_list = to_list(
                 feature_names, element_type=str, arg_name="feature_names"
             )
-            _ensure_columns_exist(column_type="feature", columns=features_list)
+            _ensure_columns_exist(
+                observations, column_type="feature", columns=features_list
+            )
 
             # ensure features and target(s) do not overlap
             shared = set(targets_list).intersection(features_list)
@@ -151,18 +140,16 @@ class Sample:
 
         self._feature_names = features_list
 
-        # make sure the index has a name
-
-        if observations_index.name is None:
-            observations = observations.rename_axis(index=Sample.IDX_OBSERVATION)
-
         # keep only the columns we need
 
         observation_columns = [*features_list, *targets_list]
         if weight_name is not None and weight_name not in observation_columns:
             observation_columns.append(weight_name)
 
-        self._observations = observations.loc[:, observation_columns]
+        # select just the columns we need to retain and tidy up the observations table
+        self._observations = _tidy_up_observations(
+            observations.loc[:, observation_columns]
+        )
 
     @property
     def index(self) -> pd.Index:
@@ -262,7 +249,7 @@ class Sample:
         subsample = copy(self)
         if iloc is None:
             if loc is None:
-                ValueError("either arg loc or arg iloc must be specified")
+                raise ValueError("either arg loc or arg iloc must be specified")
             else:
                 subsample._observations = self._observations.loc[loc, :]
         elif loc is None:
@@ -325,3 +312,46 @@ class Sample:
 
 
 __tracker.validate()
+
+#
+# auxiliary functions
+#
+
+
+def _ensure_columns_exist(
+    observations: pd.DataFrame, column_type: str, columns: List[str]
+) -> None:
+    # check if all provided feature names actually exist in the observations df
+    available_columns: pd.Index = observations.columns
+    missing_columns = {name for name in columns if name not in available_columns}
+    if missing_columns:
+        raise KeyError(
+            f"observations table is missing {column_type} columns {missing_columns}"
+        )
+
+
+def _tidy_up_observations(observations: pd.DataFrame) -> pd.DataFrame:
+    # ensure all column names are native Python strings
+    name_types = {type(name) for name in observations.columns}
+    name_types.discard(str)
+    invalid_name_types = [
+        name_type
+        for name_type in name_types
+        if not np.issubdtype(name_type, np.character)
+    ]
+    if invalid_name_types:
+        # not all names are strings
+        raise TypeError(
+            "all column names in arg observations must be strings, but included: "
+            + ", ".join(t.__qualname__ for t in invalid_name_types)
+        )
+
+    # convert numpy string types to native Python strings
+    if name_types:
+        observations = observations.set_axis(observations.columns.astype(str), axis=1)
+
+    # ensure the index has a name
+    if observations.index.name is None:
+        observations = observations.rename_axis(index=Sample.IDX_OBSERVATION)
+
+    return observations